runDestroy

+++ src/http/post-signup-delete/index.js
@@ -4,19 +4,27 @@ let data = require('@begin/data')
 
 exports.handler = async function destroy (req) {
   // let key = arc.http.helpers.bodyParser(req).key // Base64 decodes + parses body
-  // console.log(key)
   
-  await data.destroy({
-    table: 'accounts',
-    key: ''
-    
-  })
+  let table = 'accounts'
+  let tableCount = await data.count({table})
+
+  let run = () => {
+    if (tableCount <= 0)
+    then
+    await data.destroy({
+      table: 'accounts',
+      key: 'Y5KxlnRZuD'
+      
+    })
+  }
+  run()
 
   return {
     statusCode: 302,
     headers: {
       location: '/',
       'cache-control': 'no-cache, no-store, must-revalidate, max-age=0, s-maxage=0'
-    }
+    },
+    
   }
 }
